# spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  description: Bla bla bla
  version: 1.0.0
  title: User and team - LozziKit
host: 'localhost:8080'
basePath: /
tags:
  - name: user
    description: Operations about user
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /users:
    post:
      tags:
        - user
      summary: Create user
      description: Create a new user
      operationId: createUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/NewUser'
      responses:
        '201':
          description: successful operation
          headers:
            Location:
              type: string
        '400':
          description: Bad request
          schema:
            type: string
        '409':
          description: Username/email exists already
    get:
      security:
        - Bearer: []
      tags:
        - user
      summary: Get users
      description: Get all the users
      operationId: getUsers
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
  '/users/{username}':
    get:
      security:
        - Bearer: []
      tags:
        - user
      summary: Get user
      description: Get a user
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Modify user
      description: Modify an existing user
      operationId: updateUser
      produces:
        - application/json
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched'
          required: true
          type: string
        - in: body
          name: body
          description: Modified user object
          required: true
          schema:
            $ref: '#/definitions/UserModified'
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          schema:
            type: string
        '409':
          description: Username/email exists already
  '/users/{username}/teams':
    get:
      security:
        - Bearer: []
      tags:
        - user
      summary: Get user
      description: Get a user
      operationId: getTeamsByUser
      produces:
        - application/json
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
        '404':
          description: User not found
  /auth:
    post:
      tags:
        - auth
      summary: Authenticate a user
      description: Authenticate a user
      operationId: authUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        '200':
          description: successful operation (Get JWT token)
          schema:
            type: string
        '400':
          description: Bad request
          schema:
            type: string
        '401':
          description: Unauthorized username or password
          schema:
            type: string
  /teams:
    post:
      tags:
        - team
      summary: Create team
      description: Create a new team
      operationId: createTeam
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created team object
          required: true
          schema:
            $ref: '#/definitions/NewTeam'
      responses:
        '201':
          description: successful operation
          headers:
            Location:
              type: string
        '400':
          description: Bad request
          schema:
            type: string
        '409':
          description:  Team name exists already
    get:
      security:
        - Bearer: []
      tags:
        - team
      summary: Get teams
      description: Get all the teams
      operationId: getTeams
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
  '/teams/{id}':
    get:
      security:
        - Bearer: []
      tags:
        - team
      summary: Get team
      description: Get a team
      operationId: getTeam
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The id of the team'
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Team'
        '404':
          description: Team not found
    put:
      tags:
        - team
      summary: Modify team (generally adding or removing users)
      description: Modify an existing team
      operationId: updateTeam
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The id of the team that needs to be fetched'
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Modified team object
          required: true
          schema:
            $ref: '#/definitions/Team'
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          schema:
            type: string
        '409':
          description: Team name already exists
definitions:
  NewUser:
    type: object
    properties:
      username:
        type: string
        minLength: 4
      firstName:
        type: string
        minLength: 4
      lastName:
        type: string
        minLength: 4
      email:
        type: string
        format: email
        pattern: '^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$'
        minLength: 6
      password:
        type: string
        format: password
        minLength: 4
    required:
      - username
      - firstName
      - lastName
      - email
      - password
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
    required:
      - id
      - username
      - firstName
      - lastName
      - email
  UserModified:
    type: object
    properties:
      firstName:
        type: string
        minLength: 4
      lastName:
        type: string
        minLength: 4
      email:
        type: string
        pattern: '^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$'
        minLength: 6
      password:
        type: string
        format: password
        minLength: 4
    required:
      - firstName
      - lastName
      - email
      - password
  Credentials:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    required:
      - username
      - password
  Team:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        minLength: 4
      members:
        type: array
        items:
          $ref: '#/definitions/User'
  NewTeam:
    type: object
    properties:
      name:
        type: string
        minLength: 4
      members:
        type: array
        items:
          $ref: '#/definitions/User'
    required:
      - name